requirements.txt
Flask==3.0.0
google-generativeai==0.3.0
Pillow>=10.0.0
opencv-python>=4.8.0
python-dotenv==1.0.0
easyocr==1.7.0
torch>=2.0.0
torchvision>=0.15.0
